% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{wk_assignments_by_id}
\alias{wk_assignments_by_id}
\title{Get Assignment by ID}
\usage{
wk_assignments_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer assignment ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing assignment data
}
\description{
Returns a specific assignment by ID.
}
\examples{
\dontrun{
assignment <- wk_assignments_by_id(123456)
}

}
