% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{wk_extract_data}
\alias{wk_extract_data}
\title{Extract Data Items}
\usage{
wk_extract_data(response)
}
\arguments{
\item{response}{An API response object}
}
\value{
The data items from the response
}
\description{
Helper function to extract just the data items from an API response,
discarding metadata and pagination info.
}
\examples{
\dontrun{
response <- wk_subjects(levels = 1)
subjects <- wk_extract_data(response)
}

}
