% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{wk_has_next_page}
\alias{wk_has_next_page}
\title{Check if More Pages Exist}
\usage{
wk_has_next_page(response)
}
\arguments{
\item{response}{An API response object}
}
\value{
Logical indicating if more pages are available
}
\description{
Helper function to check if there are more pages available.
}
\examples{
\dontrun{
response <- wk_subjects(levels = 1)
if (wk_has_next_page(response)) {
  # Fetch next page
}
}

}
