% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_progressions.R
\name{wk_level_progressions_by_id}
\alias{wk_level_progressions_by_id}
\title{Get Level Progression by ID}
\usage{
wk_level_progressions_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer level progression ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing level progression data
}
\description{
Returns a specific level progression by ID.
}
\examples{
\dontrun{
progression <- wk_level_progressions_by_id(123456)
print(progression$data$level)
}

}
