% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resets.R
\name{wk_resets_by_id}
\alias{wk_resets_by_id}
\title{Get Reset by ID}
\usage{
wk_resets_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer reset ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing reset data
}
\description{
Returns a specific reset by ID.
}
\examples{
\dontrun{
reset <- wk_resets_by_id(123456)
}

}
