% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_statistics.R
\name{wk_review_statistics_by_id}
\alias{wk_review_statistics_by_id}
\title{Get Review Statistic by ID}
\usage{
wk_review_statistics_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer review statistic ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing review statistic data including:
  \itemize{
    \item meaning_correct: Number of correct meaning answers
    \item meaning_incorrect: Number of incorrect meaning answers
    \item meaning_current_streak: Current streak of correct meaning answers
    \item meaning_max_streak: Maximum streak of correct meaning answers
    \item reading_correct: Number of correct reading answers
    \item reading_incorrect: Number of incorrect reading answers
    \item reading_current_streak: Current streak of correct reading answers
    \item reading_max_streak: Maximum streak of correct reading answers
    \item percentage_correct: Overall percentage of correct answers
  }
}
\description{
Returns a specific review statistic by ID.
}
\examples{
\dontrun{
stats <- wk_review_statistics_by_id(123456)
print(stats$data$percentage_correct)
}

}
