% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{wk_reviews_by_id}
\alias{wk_reviews_by_id}
\title{Get Review by ID}
\usage{
wk_reviews_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer review ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing review data
}
\description{
Returns a specific review by ID.
}
\examples{
\dontrun{
review <- wk_reviews_by_id(123456)
}

}
