% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaced_repetition_systems.R
\name{wk_spaced_repetition_systems_by_id}
\alias{wk_spaced_repetition_systems_by_id}
\title{Get Spaced Repetition System by ID}
\usage{
wk_spaced_repetition_systems_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer SRS ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing SRS data
}
\description{
Returns a specific spaced repetition system by ID.
}
\examples{
\dontrun{
srs <- wk_spaced_repetition_systems_by_id(1)
print(srs$data$name)
print(length(srs$data$stages))
}

}
