% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_materials.R
\name{wk_study_materials_by_id}
\alias{wk_study_materials_by_id}
\title{Get Study Material by ID}
\usage{
wk_study_materials_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer study material ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing study material data
}
\description{
Returns a specific study material by ID.
}
\examples{
\dontrun{
material <- wk_study_materials_by_id(123456)
}

}
