% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{wk_user_update}
\alias{wk_user_update}
\title{Update User Preferences}
\usage{
wk_user_update(preferences, api_token = NULL)
}
\arguments{
\item{preferences}{A named list containing preference settings to update.}

\item{api_token}{Character string with your WaniKani API token. If NULL, uses
the WANIKANI_API_TOKEN environment variable.}
}
\value{
A list containing updated user information.
}
\description{
Updates the user's preferences.
}
\examples{
\dontrun{
# Update user preferences
prefs <- list(
  lessons_autoplay_audio = TRUE,
  lessons_batch_size = 5
)
user <- wk_user_update(preferences = prefs)
}

}
