% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice_actors.R
\name{wk_voice_actors}
\alias{wk_voice_actors}
\title{Get Voice Actors}
\usage{
wk_voice_actors(ids = NULL, updated_after = NULL, api_token = NULL)
}
\arguments{
\item{ids}{Optional integer vector of voice actor IDs}

\item{updated_after}{Optional character timestamp (ISO 8601)}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing voice actor data and pagination information. Each
  voice actor includes:
  \itemize{
    \item name: Name of the voice actor
    \item gender: Gender of the voice actor
    \item description: Description of the voice actor
  }
}
\description{
Returns a collection of voice actors filtered by query parameters. Voice actors
are used for vocabulary audio.
}
\examples{
\dontrun{
# Get all voice actors
voice_actors <- wk_voice_actors()
}

}
