% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice_actors.R
\name{wk_voice_actors_by_id}
\alias{wk_voice_actors_by_id}
\title{Get Voice Actor by ID}
\usage{
wk_voice_actors_by_id(id, api_token = NULL)
}
\arguments{
\item{id}{Integer voice actor ID}

\item{api_token}{Character string with your WaniKani API token}
}
\value{
A list containing voice actor data
}
\description{
Returns a specific voice actor by ID.
}
\examples{
\dontrun{
voice_actor <- wk_voice_actors_by_id(1)
print(voice_actor$data$name)
}

}
